runner: create TaskRunner interface
