+++ plugins/plugin.go
@@ -3,6 +3,8 @@ package plugins
 import (
 	"github.com/x1unix/gilbert/logging"
 	"github.com/x1unix/gilbert/manifest"
+	"github.com/x1unix/gilbert/runner"
+	"github.com/x1unix/gilbert/runner/job"
 	"github.com/x1unix/gilbert/scope"
 )
 
@@ -15,5 +17,5 @@ type PluginFactory func(*scope.Scope, manifest.RawParams, logging.Logger) (Plugi
 // Plugin represents Gilbert's plugin
 type Plugin interface {
 	// Call calls a plugin
-	Call() error
+	Call(ctx *job.RunContext, r runner.TaskRunner) error
 }
